requirements.txt
flask
psycopg2-binary
flask_sqlalchemy
flask_login
requests
